fixed error with stock prices dropping below $.01

+++ ticker.h
@@ -7,7 +7,7 @@
 
 struct company {
     char symbol[6];
-    size_t cents;
+    int cents;
     char *name;
     int status; // 1 is add 0 is no symbol and -1 is subtract 
 };
